Errors fixed
