+++ .idea/php.xml
@@ -1,4 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="PhpIncludePathManager">
+    <include_path>
+      <path value="$PROJECT_DIR$/Blog/vendor/fakerphp/faker" />
+      <path value="$PROJECT_DIR$/Blog/vendor/symfony/deprecation-contracts" />
+      <path value="$PROJECT_DIR$/Blog/vendor/composer" />
+      <path value="$PROJECT_DIR$/Blog/vendor/psr/container" />
+    </include_path>
+  </component>
   <component name="PhpProjectSharedConfiguration" php_language_level="7.4" />
 </project>
